pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rdl2arch"
version = "0.1.0"
description = "SystemRDL to ARCH HDL register-block generator"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [{ name = "rdl2arch contributors" }]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "systemrdl-compiler~=1.32",
    "Jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "mypy>=1",
]

[project.scripts]
rdl2arch = "rdl2arch.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
rdl2arch = ["templates/*.j2"]
